Huge UI overhaul
The snap scrolling was causing too many issues on mobile, so we've totally overhauled the layout. Now, on desktop, instead of scrolling, we just have tabs at the top of the screen to navigate to each section. On mobile, however, there is now a bottom navigation div, with arrows to move between the sections.